move collection and scheduledItems call to router, remove duplicate main test
