+++ src/routes/collection.integration.ts
@@ -3,11 +3,9 @@ import config from '../config';
 import request from 'supertest';
 import { app } from '../main';
 import { brazeCollectionsFixture, graphCollectionFixture } from './fixture';
-import * as collection from './collections'
 import { client } from '../graphql/client-api-proxy';
 
 describe(`get collection test`, () => {
-  const requestAgent = request.agent(app);
   beforeEach(() => {
     jest.clearAllMocks();
   });
@@ -16,7 +14,7 @@ describe(`get collection test`, () => {
     const testSlug = 'the-world-as-explained-by-pop-culture';
     // spying on the getStories function to make it return a mock response
     jest.spyOn(client,'query').mockResolvedValue(graphCollectionFixture as any);
-    const response = await requestAgent.get(
+    const response = await request(app).get(
       `/collection/${testSlug}?apikey=${config.aws.brazeApiKey}`
     );
     expect(response.statusCode).equals(200);
@@ -25,22 +23,33 @@ describe(`get collection test`, () => {
     expect(response.headers['cache-control']).to.equal('public, max-age=120');
   });
 
-  it('should return 500 if invalid api key is provided ', async () => {
-    const serverError = {
+  it('should return 500 if collection is not found', async () => {
+    const notFoundGraphError = {
       "errors": [
         {
-          "message": "something went wrong",
+          "message": "Error - Not Found: not-found",
           "path": [
             "getCollectionBySlug"
           ],
           "extensions": {
-            "code": "INTERNAL_SERVER_ERROR",
+            "code": "NOT_FOUND",
             "serviceName": "collection"
           }
         }
       ],
+      "data": {
+        "getCollectionBySlug": null
+      }
     };
-    const response = await requestAgent.get(
+    jest.spyOn(client,'query').mockResolvedValue(notFoundGraphError as any);
+    const response = await request(app).get(
+      `/collection/not-found-slug?apikey=${config.aws.brazeApiKey}`
+    );
+    expect(response.statusCode).equals(500);
+  });
+
+  it('should return 500 if invalid api key is provided ', async () => {
+    const response = await request(app).get(
       `/collection/this-is-test-slug?apikey=invalid-api-key`
     );
     expect(response.statusCode).equals(500);
